renamed adminpanel

+++ src/components/TasksDashboard/TasksDashboard.tsx
@@ -1,6 +1,5 @@
 import { useEffect, useState } from 'react';
 import axios from 'axios';
-import { ETaskStatus } from './../../enums/ETaskStatus/ETaskStatus';
 import { ITask } from '../../types/ITask/ITask';
 import TasksContainer from '../TasksContainer/TasksContainer';
 import EditTaskModal from '../EditTask/EditTaskModal';
@@ -10,10 +9,10 @@ import './TasksDashboard.css';
 
 const TasksDashboard = () => {
   const [tasksFromBack, setTasksFromBack] = useState<{
-    toDo: Array<ITask>,
-    inprogress: Array<ITask>,
-    review: Array<ITask>,
-    done: Array<ITask>,
+    toDo: ITask[];
+    inprogress: ITask[];
+    review: ITask[];
+    done: ITask[];
   }>({
     toDo: [],
     inprogress: [],
@@ -30,16 +29,12 @@ const TasksDashboard = () => {
   }, []);
 
   const getData = async () => {
+    if (!token) return;
     try {
       const res = await axios.post(
         'https://localhost:7095/api/Issue/GetAllIssues',
-        {
-          text: '',
-          userIds: [''],
-        },
-        {
-          headers: { Authorization: `Bearer ${token}` },
-        }
+        { text: '', userIds: [''] },
+        { headers: { Authorization: `Bearer ${token}` } }
       );
       setTasksFromBack(res.data.data);
     } catch (error) {
@@ -48,13 +43,13 @@ const TasksDashboard = () => {
   };
 
   const handleDeleteTask = async (id: number) => {
+    if (!token) return;
     try {
       const res = await axios.delete(`https://localhost:7095/api/Issue/${id}`, {
         headers: { Authorization: `Bearer ${token}` },
       });
-
       if (res.status === 200) {
-        getData();
+        await getData();
       }
     } catch (error) {
       console.error('Error deleting task:', error);
@@ -66,45 +61,50 @@ const TasksDashboard = () => {
   };
 
   return (
-    <div className='full-page'>
-      <div className='dashboard-page'>
-        <div className='top-section'>
-          <h2 className='title'>Tasks</h2>
-          <button className='task-btn' onClick={() => setIsNewTaskModalOpen(true)}>
+    <div className="full-page">
+      <div className="dashboard-page">
+        <div className="top-section">
+          <h2 className="title">Tasks</h2>
+          <button
+            className="task-btn"
+            onClick={() => {
+              setIsNewTaskModalOpen(true);
+            }}
+          >
             New Task ✚
           </button>
         </div>
 
-        <div className='tasks-list'>
+        <div className="tasks-list">
           <TasksContainer
-            name='To Do'
+            name="To Do"
             tasks={tasksFromBack.toDo}
-            backgroundColor='rgba(188, 206, 234 , 0.3)'
-            borderColor='rgb(83, 135, 219)'
+            backgroundColor="rgba(188, 206, 234 , 0.3)"
+            borderColor="rgb(83, 135, 219)"
             onDelete={handleDeleteTask}
             setTaskToEdit={setTaskToEdit}
           />
           <TasksContainer
-            name='In Progress'
+            name="In Progress"
             tasks={tasksFromBack.inprogress}
-            backgroundColor='rgba(245, 216, 169,0.3)'
-            borderColor='rgb(221, 173, 97)'
+            backgroundColor="rgba(245, 216, 169,0.3)"
+            borderColor="rgb(221, 173, 97)"
             onDelete={handleDeleteTask}
             setTaskToEdit={setTaskToEdit}
           />
           <TasksContainer
-            name='Review'
+            name="Review"
             tasks={tasksFromBack.review}
-            backgroundColor='rgba(245, 176, 169, 0.3)'
-            borderColor='rgb(227, 138, 128)'
+            backgroundColor="rgba(245, 176, 169, 0.3)"
+            borderColor="rgb(227, 138, 128)"
             onDelete={handleDeleteTask}
             setTaskToEdit={setTaskToEdit}
           />
           <TasksContainer
-            name='Done'
+            name="Done"
             tasks={tasksFromBack.done}
-            backgroundColor='rgba(186, 216, 156, 0.3)'
-            borderColor='rgb(149, 186, 112)'
+            backgroundColor="rgba(186, 216, 156, 0.3)"
+            borderColor="rgb(149, 186, 112)"
             onDelete={handleDeleteTask}
             setTaskToEdit={setTaskToEdit}
           />
@@ -125,7 +125,7 @@ const TasksDashboard = () => {
       {isNewTaskModalOpen && (
         <NewTaskModal
           onClose={() => setIsNewTaskModalOpen(false)}
-          onTaskCreated={async () => {
+          onTaskCreated={async (newTask) => {
             await getData();
             setIsNewTaskModalOpen(false);
           }}
